Tweak CSS, make "Get Started" go somewhere
Made the navbar CSS more general so that we can more easily add new
items to it when needed on both left and right.

Also had Get Started links go to instances.mastodon.xyz temporarily,
per discussion in #landing-page on Discord that we should at least
have the Call-to-Action do something while the page is live. Once we
build our own instance-suggester, we will change this link to that.