Remove othering: user → people
Let’s try not to refer to people as people whenever possible. It removes the othering (“users” creates a different group to us as “developers” and “designers”). Sometimes it might be impossible to avoid using the term (although you’d be surprised how often you can replace “user“ with “person” and not lose any meaning). Also, as developers and designers, we are also the users of Mastodon so “user” is really more a role than an identity. When we’re talking about people, let’s try to call them people whenever possible and perhaps only use “user“ when referring to a person playing that role at a specific time (when we need to expressly differentiate the role from other roles).

(Also, as evidenced in this case, the title reads better and sounds more human.) :)