feat: add support for ?lang search parameter
This allows users to give links such as joinmastodon.org/?lang=fr to
their friends, who on visit will be seeing, in this case, French
version of the site, but users can also supply any other locale there:
if locale is not found, it will fallback to auto-detection or English.

We could have also added a copy of index.html page for every locale,
but I'm not sure how to do this, so will leave it as is.