Strengthen the server covenant
Explicitly reference abuse of external origin. A server that has
a code of conduct and a small number of users may never feel the
need to moderate abusive behavior. But if such a server fails
to suspend the worst routine origins of abuse in the fediverse,
it ultimately fails its users.

In practice, new users signing up, especially vulnerable folks,
may be immediately targeted by abusive actors from other servers.
Several instances of exactly this happening have been
well-documented by now.

Since this extent of moderation responsibility is non-intuitive
and non-obvious, it is important to be explicit.